feat: add context for the main app theme
